Implemented sauna commands

+++ MijnThuis.Contracts/Sauna/GetSaunaOverviewResponse.cs
@@ -5,5 +5,5 @@ public class GetSaunaOverviewResponse
     public string State { get; set; }
     public int InsideTemperature { get; set; }
     public int OutsideTemperature { get; set; }
-    public int Power { get; set; }
+    public decimal Power { get; set; }
 }
